Add support for partition image undercloud.qcow2
We are trying out a new method for building the undercloud image
in RDO[1]. This will make it impossible to drift from upstream
content in the overcloud images. However, it means the
undercloud.qcow2 that is produced is no longer a whole disk image.

This change adds support to use an undercloud image that is
created this way by piggybacking on the devmode code that already
supports this.

It would be good to follow this up with some clean-up of the two
use cases to address the similarities in a cleaner fashion.

[1] https://review.rdoproject.org/r/4013
Change-Id: I5878ddb9f82684d775f9daa66fea434d58f9609c