Change initrd extraction method for partition undercloud
There is a strange issue with the new image building process in
RDO where the host initramfs is ending up in the undercloud.qcow2,
but the host kernel is not. This initramfs does not match the
kernel from the CentOS cloud image used as a base, but instead is
for a newer kernel. The virt-get-kernel command just takes the
newest kernel and initramfs it can find in /boot/, which in this
case ends up with a mismatch.

Since the new image building process is just a converted
overcloud-full image, we can just copy-out the DIB extracted kernel
and initramfs from the undercloud to use for booting it. DIB seems
to be a bit smarter about how it extracts the kernel and initramfs,
so we do not get the same mismatch.

Change-Id: I66f6f168417377f313c003be70ee708310371f35