Allow permissive access to non_root_user files
Openvswitch ports cannot be connected to unprivileged libvirt
guests, so the only way to use OVS on the undercloud host at
the moment is to run the guests with privileges.

In this case, the root user is used to launch the VMs, and the
new option added in this patch can be used to give access to
the image files owned by the non_root_user.

Change-Id: Ife7f10df3d89798d3c5753f5ddc9a06eda5ef8d9