Add the ARA Ansible callback generate a static playbook reports
ARA [1] records ansible-playbook runs.
It provides a CLI and web interface to browse detailed and organized
information about the results of those playbook runs.

This commit makes it so ARA is installed and set up as a callback
and then generates a static report that can later be picked up
by the collect-logs role.

[1]: http://ara.readthedocs.io/en/latest/faq.html#what-is-ara

Change-Id: If65bbd0259229c0db684ea628814f90d8cde12e8