ensure the volume pool can be marked autostart in all cases
In some cases after rerunning the workflow it is possible to
get into a state where the oooq pool is running but not
autostarted.  In this case the pool_check will pass and pool
autostart will fail.  This should provide the steps required
for autostart to work in either case.

Closes-Bug: #1657232

Change-Id: Ieba0caa40290c9e03740cef337c6a3f3fa04db30