Replace ansible_user with ssh_user
ansible_user is currently used both as the builtin variable
and a regular variable. Since its value changes multiple times
over the course of a deployment, determining its value at any given
point in time can be difficult. In addition, overriding its value
in order to connect to the virthost with a different user has
unexpected results for the same reason.

ansible_user is therefore replaced  with ssh_user, which is never
modified from its default in common defaults, and can be safely
overridden by the user.

Change-Id: Id9bc7c2a3b11140508ef0911de317a4f1ece05dd