ensure that tripleo-quickstart volume pool is defined
I wanted to suggest an alternative to the solution proposed in
https://review.openstack.org/#/c/421467/, but I was too late.  This
change reduces the three tasks added in #421467 to a single task,
since defining a volume pool is idempotent.  We don't need to check if
is already defined or not first.

This also corrects the fact that the previous changes are checking for
the existence of an XML file that libvirt doesn't actually care about.

Change-Id: I726b89e4bbda18c9648ce52da12b2728a232e00c