Add repo configuration to release configs
The newly created repo-setup role in tripleo-quickstart-extras[1],
allows for yaml based configuration of repos. This is what defines
the release when not using an image, or when creating an image.

As such, we should include the repo configuration in each release
config.

Much of the centosci directory is left untouched as I am unsure
how all of the various configs there are being used.

Change-Id: Ic7acd1d333084089337d6b74c11fde190dbb8676