break out general_config into config and topology
This change takes the overcloud_nodes variable and defines that
under config/nodes/foo.yml. This depend on the change
to quickstart-extras where a definition of topology (nodes)
will be created. Pass the topology in an explicit variable vs.
using extra_args.

The original configuration files in config/general_config/ will
continue to work but are now deprecated.
The featureset files that define standard set of feature to test are
also added, and should correspond to test configuration for jobs that
are typically found upstream

Change-Id: Id6dd3f71b33f76a0401fe8a84080fe80de0dbb69
Co-Authored-By: Gabriele Cerami <gcerami@redhat.com>