Use md5sum -b on images
according to md5sum man page

-b, --binary
      read in binary mode

The  sums  are  computed  as  described in RFC 1321.  When checking, the input
should be a former output of this program.  The default mode is to print a line
with checksum, a space, a character indicating input mode ('*' for binary, ' '
for text or where binary is insignificant), and name for each FILE.

the ramification is that this can be different if read binary is different from
from normal read on a given platform.  The binary option is better if you know
what you are trying to check is a binary blob.

Change-Id: Ib1490feefe4e026b3834a40ae9072a82af1379ee