[convert-image] Add ability to override the directory used
It is dangerous (if not impossible) to override the global
working_dir variable, because it is used in many different places.

Instead we create a new variable internal to this role which
defaults to the global working_dir, but can be overridden solely
for the single role.

This is needed for the build-images role so that we can build the
images in a different directory than the working_dir.

Change-Id: I6177e140c8585cb5d1b994e681d9f08acfaf0cef