main script: add zuul-clone default path for upstream ci
the esisting block in charge of cloning the correct commit of extras in
case of upstreeam depends-On relies on a environment variable that was
missing.
Adding ZUUL_CLONER variable at the start of the script with the default
devstack-gate location, allows quickstart to clone the correct
dependent commit in upstream.

Change-Id: Idefcd501e22f961dbd07c826e37816861888abe2