Add option to pass a config file with full path to collect-logs.sh
The current script accepts only a single yaml file  in one
directory as a config that can contain documentation generation
related variables.
This review adds the option to pass the full path to the config
file so that a custom file in any directory can be used for
doc generation options.

Closes-Bug: #1661657
Change-Id: I76bc8f560a5b06339c91609dcc781d376e820d7d