define non_root_group explicitly
previously, we used set_fact to set the non_root_group variable but
this is problematic: the libvirt/ playbooks are expected to be
runnable *without* also running the provision/ playbooks, in which
case non_root_group is not set, causing ansible to fail.

this change makes non_root_group explicit and adds it to
roles/common/defaults/main.yml.

Change-Id: I5a174a303c980569fad31f7d7fa8e846105a2291