Enable quickstart to work with clean base OS images.
Enables a deployment using clean base os RHEL/CentOS images.
A clean base os image would be the standard cloud images for
CentOS and RHEL.  The normal path for quickstart is to use
a prebuilt undercloud image with all the packages preinstalled.

* undercloud_image_url points to a base os image and includes
a md5sum for that image in the config.

* update fetch.yml to handle base os images

* create a baseos convert image script in convert-image that
does not install openstack packages.

* create a new variable call baseos_as_undercloud variable

Change-Id: Ia2394943c12f3bc6ac29235ca741e0eb2dd17db3