Ignore libvirt start if running chrooted
The use case is when running oooq from a
privileged container. For examle, running in
Centos container on the Ubuntu virthost.
Services can't be restarted in such a container
unless you run it as a VM (with init as pid 1).
So oooq can't pass the libvirt startup task and
proceed with undercloud deployment. This is
controlled by a new var 'chrooted'. Defaults to
False.

Partial-bug: #1676373
Change-Id: If221a6a8bda496862da6249c46c2bedd955b6eac
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>