Allow custom undercloud initrd/vmlinuz
Preserve default behavior, which is extract
those from the overcloud image.

The example use case is when running oooq from
a priv container. Like running in Centos container
on the Ubuntu virthost. The virt-copy-out may be
working fine on the host OS and failing within the
container with an error like:
"'/dev/urandom' is not a supported random source".
So those images may be prepared as well at the host
OS and reused as external ones.

Partial-bug: #1676373
Change-Id: I19db238b4966a59675e42983daccffe3c437019c
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>