Instack.json for undercloud VM image
Do not try to copy the instackenv.json configuration
file to the undercloud VM image, if it wasn't generated
in the overcloud setup role.

This may be the case if there are no overcloud nodes
defined and the overcloud setup role is missing in
the play, f.e. the containerized undercloud use case.

Partial-bug: #1676373
Change-Id: I7a5a564aacdfa2153ad727baf0b5259cb723ec69
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>