Fix for quickstart image building
In 57443434af78565cb7850886383f04b5e3ab4a08 we introduced a check
for overcloud_as_undercloud to the convert-image role in order to
make it generic enough to handle converting a baseOS image. This
had the unintended consequence of breaking image building where
we do not set this variable.

Instead, we can just check if the overcloud-full image exists, and
only convert it in that case.

Change-Id: Ica9f55b970db12d2fd98ae6a38a0ee99b171ae1d