Switch tripleo-ci to use trunk.rdoproject.org repos
This is a similar patch to https://review.openstack.org/452174
We still have this custom code for libvirt devmode, so we need to
change the repo here too. This is a good reminder that we need to
remove this custom code and replace it with a call to the repo-setup
role. This patch addresses the major breakage, but I will post a
follow-up to refactor this.

Change-Id: I4d772756a652448203a74f3ca75cacf4ec0d42b6
Closes-Bug: 1678218