repo setup can not occur until users are setup
In an attempt to ensure yum repos were in place for any possible package
installs the include statement was added agressively to the beginning
of the workflow.  This passed local testing because the user was present
even with --teardown all initially.  In fact we have to move it after
some of the provisioning steps.  This has been confirmed to work with
a fresh install of RHEL 7.3 on the virthost w/ minimal packages.

Change-Id: I40adea3349c2fc0a2565b6e8ff1718786b9decc5