Stop using CentOS buildlogs CDN
We have had a bug open for a month[1] to fix the syncing from RDO
to the CDN, but it has still not been addressed.

This leads to a user getting stale content in the images and from
the DLRN server. Since we regularly need to purge old repos from
the DLRN server to save space, the stale content actually means we
are missing packages.

This will unfortunately mean much slower downloads for the images,
because the rdpoproject server does not have much bandwidth.
However, it is a much worse user experience to have the install
fail on missing packages.

I have intentionally left the liberty files untouched, as they
should be removed in a different commit.

[1] https://bugs.centos.org/view.php?id=12994

Change-Id: I698f0d161a3cd8fb26cd7eb4d9f43ccc15c8b9a1