add a config for building images
ATM build images is using the default configuration of
config/general_config/minimal.yml which is not appropriate
at all for building images.  Build images is not deploying
an undercloud or overcloud w/ options, it's building an image.

Most of the image build config is default, however some additional
logging is required.

Change-Id: Ibaa1f01942719d12cec1fa9b9c78980c879e80cd