Fix depends-on to use the cloned repo
After we rename the newly checked out tripleo-quickstart-gate-repo
directory to tripleo-quickstart, the script does not change to the newly
renamed directory. The file system reference remained the same (old
directory) so the change is not used.

This commit fixes the behavior by changing the working directory to the
gated tripleo-quickstart directory.

Change-Id: I82c6e93e823a06a8f7621a074c087be0046882aa