Make libselinux-python workaround less intrusive
We currently use the system python interpreter rather than the one
from the virtualenv in order to workaround the fact that the
SELinux python bindings are not available on pypi.

This is a problem if we need to use any python modules from the
virtualenv, since they will not be available to the system python
path.

Instead, we copy the python bindings from the system's site packages
(if they exist). We add a $LIBSELINUX_PYTHON_PATH var that can be
used to override this path if it is different on different systems.
By default it is set to the path in CentOS7:
lib64/python2.7/site-packages

Change-Id: Ia99a4ed16aec5d42291a2887399808c15956dc73