Use dynamic include for conditional/tagged tasks
Ansible 2.1 tries to predict static/dynamic cases for
include. It sometimes fails to do so for tagged or conditional
include commands. This makes all included tags/tasks to be played,
but skipped anyway, even though it was asked to be omitted.

Help ansible to predict dynamic cases by explicit 'static: no'
directives put for 'when:/tags:' include commands. This speeds up
deployment a lot.

Change-Id: I3daf21e45223c5165abee74e92e1395dee5f6fe9
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>