baseos-undercloud release configuration for all supported releases.
Using a baseos-undercloud configuration file replaces the quickstart
prebuilt undercloud in lieu of a base CentOS 7 qcow2.  The undercloud
is installed without any preinstalled rpms on the system.  The remaining
workflow remains the same with regards to the overcloud.

This kind of change should require a change to our gates
https://review.rdoproject.org/r/#/c/6694/

Change-Id: I51376126583448b2015004a6b65c780b6794e61d