Use repo-setup when injecting repos
We still have one-off code from when we first implemented
devmode in the task that injects repos into downloaded images.

This code works only for the master-tripleo-ci release we use
for gating. In order to allow using other releases (like
master-consistent for promote), we need to instead use the
generic repo-setup role.

Closes-Bug: 1674799
Change-Id: Id7f22e5e386df0b80f1a4abd104e644f47bd14d3