Transpose the featureset documentation table
The featureset table is growing too large as we add new featuresets, and
it's hard to solve conflicts.

By using featuresets on lines rather than columns, we can expand the
table more naturally, and the resolving conflicts will be much easier
since we can manipulate whole lines, and source code diffs work in a
line-based fashion.

Change-Id: I637da5d26a1242c194910d6b2938d731e264134b