formalize the notion of an environment configuration
Create the structure and notion of environment configuration
in tripleo-quickstart.  Environmental configuration would be
defined as any configuration that does not change featureset
values or node topology.  This is for the naunces that are
required to run tripleo in different host environments whether
that be rdo-ci, tripleo-ci, ovb in a specific cloud etc.

Change-Id: I4cf2e1cac0325837deeefa889a3699e0db0ec7d8