qemu://system fixes
When a libvirt uri of qemu:///system is used, the root user must
be used to control power on VMs, since they are launched with
privilege. Adding the virt power key to root's authorized keys
gives the vbmc the ability to interact with privileged vms.

In addition, libguestfs requires the direct flag in order to
function when the root user is used (ssh_user: root).

Change-Id: If13befee037c76af896b14b7a085910c3d917bbe