Fix the depends-on gating for basic.sh
The gate-quickstart role uninstalls the 'tripleo-quickstart-extras'
package from the virtual env if it finds a change for it, while
modifying the quickstart-extras-requirements.txt to get it reinstalled
from the locally checked out version with the next run. That run
requires the --bootstrap option to rerun the pip commands and reinstall.
This was missing from basic.sh.

The previous workaround is not needed which manually installed the
repo. Removing the workaround fixes a bug that broke devmode.sh.

Closes-bug: #1696532
Change-Id: I65a1b286a03e9f039c75b930e3eca5ceedeb0fd6