Enable and document tags for all tasks
This commit modifies the main playbook, quickstart.yml, to have a tag
for each task declared (where it was missing).
It also adds documentation on how to use specific tags to interact with
the setup from a developer/user point of view.
Docs is still WIP, needs to be completed with extras integration and
examples (See Depends-On).

Depends-On: https://review.openstack.org/#/c/473491/
Change-Id: Ibb9900c4f8cf7c7028619ca29d288b67fb5ddc7d