Fix master-tripleo-ci release config
We no longer get both vmlinuz and kernel files for the
ironic-python-agent in tripleo. The vmlinuz was always redundant,
so this does not have any impact other than failing when we try
to inject it into the undercloud. This can be easily resolved by
just not trying to inject the missing image.

Change-Id: I706bb7d89fa4b12bab887e41cb2d6e1001c7058d