Move common multinode role setup from tripleo-ci
This configuration was previously in tripleo-ci in the
testenv/multinode.yml file, but it is not specific to tripleo-ci.

This would be nice to add to the common multinode config and only
override where needed, but our current code does not allow for
overriding what is in the common config.

Change-Id: Ia042fc578ade2c073504ceb117b1da3ff0a4e051