Default update_images to false
When removing the hardcoded repo setup from devmode, we actually
changed the behavior of when we run repo setup on the overcloud
images.

Instead, we should keep the behavior the same as before. Rather than
adding "update_images: true" to all tripleo-ci configs, we just make
the behavior trigger on either devmode or update_images vars.

Change-Id: Ic08cad28d3f6d33c3aaa35903f8013f49886712b