Use containers mirrored from rdoproject.org build/promote
We can not use the rdoproject.org registry directly, because there
is no infra mirror for it. Instead we will mirror the container images
to dockerhub when the periodic job passes. The first batch of containers
produced this way is already uploaded, and newer than the ones we have
manually built in the tripleoupstream namespace on dockerhub. As such,
we can switch to using the automated container builds.

Change-Id: I2339a72bb265470c26435e2f7401a81aff4fc2ef