pm_addr in instack should not be localhost
Currently when oooq deploy the undercloud, the instack.json file
has the following entries:
...
      "name": "compute-0",
      "pm_password": "password",
      "pm_type": "ipmi",
      "pm_user": "admin",
      "pm_addr": "127.0.0.1",
      "pm_port": "6233",
      "mac": [
        "00:a2:18:01:33:cf"
      ],
...

While this works for ironic on the undercloud, it is problematic
as soon as we use the instack.json to configure fencing for the
overcloud. Let's replace 127.0.0.1 with the ip used by the undercloud.
This way ironic on the undercloud keeps working and the fencing
configuration that gets generated via 'overcloud generate fencing
instack.json' is correct.

Change-Id: I0727a9224b5e9eaafaa374ecfbcd4fb9ccb52d58
Closes-Bug: #1656878