Fix clouds.yaml generation
Previously the os_tenant_name was being used to generate the
project_name in clouds.yaml. But if your credentials environment did not
include an OS_TENANT_NAME declaration but did include an OS_PROJECT_NAME,
this could result in a bad clouds.yaml and auth failures. This change
adds the tenant_name in the clouds.yaml as well as setting project_name
to use os_project_name while still falling back to the previous
os_tenant_name.  Additionally since this is a v2.0/v3 issue, let's
support properly defining the identity api version if defined.

Change-Id: Iea728eabc1bc00facb3f208a69dbafaa630912b2
Closes-Bug: #1723669