Get DLRN hash before setting up repos and images
When deploying with containerized overcloud and updating
containers prior to deploy, the DLRN hash of the images
used to install should match that of the containers' tags.

This review adds steps to get the DLRN hash and set the
images location, repos baseurl and containers tags.
It also updates the config/release files to
use the path to the DLRN hash related files.

Change-Id: Id8f0e4e2af57e8840d0aeb7eb9b9620e48d05279