Make provision/remote role safe to include statically
Currently including the provision/remote role without intending to
actually use it will break if the {{ virt_host_key }}.pub file does
not exist.

Ideally, we would be able to include it dynamically so that the
tasks are not evaluated at compile time, but instead only if they
are run. This would require a pretty large refactor though, because
playbooks can not be included dynamically.

Instead, we can make sure to have a default for the lookup so that
it will not fail to evaluate. The task would still fail to run if
the lookup falls back to the default, so we wont hide issues with
the key not existing with this change.

Change-Id: I347de079ada7c7b82f2f7affafa97f95c8029a3f