Disable tempest swift on scenario004
We don't have swift on scenario004, but Ceph RGW.
This patch is a technical debt, python-tempestconf should be smart
enough to know that Swift isn't actually enabled in the deployment even
if there is an object service in Keystone catalog.
It will be fixed in tempestconf, but in the meantime here to get a
promotion in CIs.

Change-Id: I512b7956dfe070e46ad2c7201f4ccad6e5024f48
Related-Bug: #1725311