Run ovb-ha with minimal services
We want to reduce the services tested by ovb-ha so we can use it to
validate overclouds can be deployed, test things that multinode can't do
now:
- HA
- net-iso
- ipv6 (will come later)
- TLS

So for that, we creating a new environment: https://review.openstack.org/522306
that will deploy a minimum list of services.
Note: for now we run the pingtest because we need https://review.openstack.org/#/c/522677/
to be promoted in RDO pipeline.

Change-Id: I58583f1699ce05b428f3b5bfdbc0bc99d1e4eecc
Depends-On: I865901f38250a13d8ea914a54072d6eb584a050d