Move set -ex to occur before activate_venv
Jobs are currently failing due to an "ansible-playbook" command not
found. It seems like the virtualenv is not being sourced properly.

Move the set -ex before activate_venv so that it not only fails fast
if it's not able to source the venv but also provides tracing to
understand what's going on.

Change-Id: Ic98679b5c62716a3abeb11abc201b67d1abf9fde