Clarify SSH server requirements prior to tripleo-quickstart deployment
On servers with more restrictive SSH server configurations, it's
important to ensure that both the root and stack users have public
key access, and that the SSH server doesn't impose any unreasonable
timeouts on the client.

This documentation update suggests in advance what the operator can
do to ensure these conditions are met.

Change-Id: Ia0ebcf30b45ed85815c8331a6aa2d5e3995ce5b9