Use current-tripleo for images in consistent configs
Images in consistent and promotion-* configs are used for local
run only, because in CI periodic jobs we always build them and
don't download.
For local run it should be safe and always existing path where
user could download image from. Both consistent and
tripleo-ci-testing links are not safe and could not to have images.
Let's use instead "current-tripleo" link, which always include
images. It's always older or equal to "consistent" so it won't
be a problem, because images are always updated in code according
to host repositories (which in our case are kept same - consistent
or tripleo-ci-testing).

Change-Id: I611aa9d23ac5f1369906fb24fd1d1198c163da2b
Closes-Bug: #1738244