Switch gate-check to THT patch
Our gate-check jobs have been broken for a while due to a
packaging issue with tripleo-ui. This switches to using a THT
patch so that we can keep getting feedback that our gates work
on a no-op change. We can switch back to the tripleo-ui patch
once the related bug is fixed.

Change-Id: I41add30cc8a71f86529b18e417a396992277fd9a
Related-Bug: 1737335