Use network templates from tripleo-heat-templates
Update quickstart ovb-ha (ipv4) jobs to use the network templates and
environments from tripleo-heat-templates instead of tripleo-ci. This
allows for having per-branch differences between the templates.

The net-iso.yaml environment from tripleo-ci is uneeded as it's
parameters are also set in network-environment.yaml, so it's usage is
removed.

Since we are not backporting all the nic config templates for ci to the
newton branch of tripleo-heat-templates, the network_isolation_args
variable is special cased to still use the templates from tripleo-ci for
the newton branch.

Change-Id: Iabad6df081555300e2c793b323acf42ea17dc5d4