Change stable repos layout
* Download TripleO packages from their "current" (e.g. current-pike)
* Download all the rest from their promoted repos (current-tripleo).

It's not realistic to deploy TripleO projects from the latest tag
promoted, some fixes have been merged in the meantime, not released and
promoted (chichen and egg, can't be released if not merged) so we really
need to deploy the TripleO projects from current.

Example: https://review.openstack.org/#/c/533380/ doesn't pass because
it's testing an old version of puppet-tripleo that doesn't contain the
fixes that we need to land https://review.openstack.org/#/c/533380/ and
fix scenario001 (voting in gate).

Therefore, let's deploy the TripleO projects from current and not from
their latest promoted version.

Change-Id: If4920c9971005ea65c1b4db11312339072ee6be9