Fix ansible warning group and host with same name
Ansible warns us that our generate inventory use the same name for
a group and a host:

  [WARNING]: Found both group and host with same name: undercloud

It doesn't make much sense to have an 'undercloud' group with just the
'undercloud' host in it. Skip group creation in the generated inventory
when the group name matches the name of the sole group host.

Change-Id: I77bf3de7b2bfc7787a07d748ba28f9ff448a71ab