Added a troubleshooting tip to clean up ~/.quickstart
I just wasted several days trying to figure out why introspection failed
in my setup with neutron authN issues in BMC node. I finally figured
that it's the same issue as bug 1706585, and the fix was to remove
~/.quickstart and start over.

Since current documentation doesn't have a page for troubleshooting tips
yet, this patch also creates one and attaches it to the bottom of the
index page.

Change-Id: I495b2063ca8b60d2fe396c776b38ea9d627831e9