In mixed version upgrade we don't want yum upgrade to be triggered.
We re-use the repo-setup role during the upgrade to create the script
we push in the UpgradeInitCommand[1] command.  We don't want "yum
upgrade -y" to be triggered at that moment.

repo_setup_run_update variable enable us to do that.  We cannot
override the repo_cmd_after because it's an extra variable (passed
using "-e") and it has the highest precedence possible in ansible[2].

[1] currently there https://review.openstack.org/#/c/541621/1/playbooks/multinode-overcloud-upgrade.yml
[2] https://docs.ansible.com/ansible/latest/playbooks_variables.html#variable-precedence-where-should-i-put-a-variable

Change-Id: I4b2f2696d366db766a5ab146cc718ccc4a15c629