Remove unnecessary parameters from featureset047.
featureset047 covers the undercloud upgrade
using tripleo-upgrade. Now that a playbook
has been enabled to run the job [0] it's not
needed to set some of the variables set when
the multinode-overcloud-upgrade-new.yml playbook
was used to run the job.
Also, overcloud validations are disabled by
setting test_ping and run_tempest to false.

[0] https://review.openstack.org/#/c/553827/

Change-Id: Id92fbbb9c25ee27fd15038ba1452b222679ae273