Add featureset051 to test keystone-only upgrades
This featureset will be used to test the simplest possible upgrade.
This will deploy the overcloud with scenario000 which is only
keystone and the basic services to make keystone work. It will
then upgrade that overcloud to release N+1.

This featureset can be used to test that all of the CI code
surrounding upgrades is working as expected, and therefore will
be perfect for using in gate jobs for the various CI repos.

Change-Id: I1b3ecc8cc7c713594c1ae8a69aa1fd1d3d7e9dfd