use minimal to verify ping/tempest
The minimal config is only used outside of the upstream CI
system.  We cover both ping and tempest upstream but not
across all the releases.  Let's use minimal.yml as the
config that should work across rleases to verify that
both ping and tempest are working correctly from a
CI standpoint.

Change-Id: I23cb01a5f753214897a0e59b4adc45e910fb3cf3