Make sure we keep using tripleo-admin as ssh user in CI
First we switched the user to tripleo-admin [1] which made CI on
updates job green, but broke Nova-based environments. We're now
switching the default back to heat-admin [2] which will break updates
job again, so we add support to tripleo-upgrade [3] to customize the
user name and set it here to the value for CI (deployed server
environments).

Fear not, when we switch to --config-download as our only supported
deployment method, tripleo-admin should become unconditionally present
in all environments, and we'll be able to converge to a single method
of logging into the overcloud nodes.

[1] I1ec66886d6d61ccf528e5e27872cf68c7a603727
[2] I0df57002b2305c1e2504c9f7a7d0c326d8ffcaf7
[3] Ib486a812aa9fd96bc753eda122cd05682d97b209

Co-Authored-By: Jose Luis Franco <jfrancoa@redhat.com>
Change-Id: I063d438fbeeded28134658aaefc6c8b7119aee43
Related-Bug: #1759845
Depends-On: I0df57002b2305c1e2504c9f7a7d0c326d8ffcaf7