Convert remaining check ovb jobs to config-download
As we move towards making config-download the default, this converts the
remaining ovb jobs in the check queue that run against master changes to
use config-download.

Once it's the default, we can come back and update these configs to not
have to explicitly use --config-download.

This also updates the feature configuration doc to indicate these ovb
jobs and the multinode jobs from the previous patch now use
config-download.

Change-Id: Ib0c0ba67e13f0978017f67dfc6095221d79b244c
implements: blueprint config-download-default