Create BuildRequires repo for master
Since Rocky, RDO is spliting dependencies repo in two different
repos, one for runtime requirement with the URL used in previous
releases and a new one with the packages that are only required
to build packages. This repo is not needed to deploy RDO but it's
needed to build packages with DLRN as done in Tripleo-CI jobs.

This patch is adding this new repository for master as it's needed
in the build-test-packages role in oooqe.

Additionally, this patch is changing URL for runtime to use
master repo.

Change-Id: I74388434e3dccfea2754569a6bc568835d49b5e2