Add ability to provision "dumb" VMs for fake nodepool
This allows the usage of the libvirt/setup/overcloud role to
provision VMs which we can use to simulate a nodepool
provisioned envrionment.

This is still a bit WIP, but the included playbook has been
successful in deploying an environment that was able to
run the tripleo-ci toci scripts on top of.

This is in a mergeable state in that it should not impact any
of our current CI which uses the libvirt/setup role. Merging
in this state will make collaborating in the upcomming sprint
around this topic easier to do asynchronously.

Change-Id: I2075e74aeb87b3268ef290181d2d5ae2672ad7a1