Fix ironic-python-agent download URLs
The ironic-python-agent is more of an undercloud part in this context,
and as such should be in sync with undercloud Ironic -- we should
fetch it from the same release as the rest of undercloud.

In case of queens-undercloud-newton-overcloud we were trying to
combinine dlrn_hash (for undercloud release, Queens) with artifacts of
overcloud release, Newton. This resulted in not just using old IPA
image, but in fact invalid URL being generated.

Change-Id: I2b0b9f81dc16344ff888e771553f6a178c0a48d8
Closes-Bug: #1764751