Rename get-undercloud-ip.sh --> get-domain-ip.sh
Context: https://trello.com/c/qgIz0V08

This commit simply renames the get-undercloud-ip.sh
script to be more generically named, as in the
libvirt-nodepool reproducer (where it's used) it is
not limited to UC.

Note: not touching the libvirt/setup/undercloud
role's version of this script (which is the same) is
intentional, and is intended to mitigate risk.

Modifying parts of TQ not utilized by libvirt nodepool
reproducer is out of scope for this commit.

Depends-On: https://review.openstack.org/#/c/565280
Change-Id: I51040db2e85296229ff98e5c691c39af8d276c0a