Remove explicit --config-download
Now that config-download is the default across tripleoclient and
tripleo-heat-templates, we can update all the CI jobs to remove the
explicit --config-download and -e
environments/config-download-environment.yaml.

Note that --deployed-server didn't actually trigger any needed behaviors
with --config-download, so where it was used, it has now been removed as
well.

Change-Id: Ic5c5df350a7eb5291d45cf8e229af575fe9bed57