Don't run full update/upgrade converge in CI to keep sane wall time
We discovered we need to put back stack updates into the converge
operations. This increases the run time of the upgrade procedure
considerably, and e.g. our voting CI job for updates might go over the
time limit. We'll need to start passing --update-plan-only to the
converge operation to keep CI on previous run times.

Change-Id: Idc5dbebc116aba2420d275aab3981186511f298e
Related-Bug: #1768586