Add config-download back into queens CI
We incorrectly dropped the config download items in
Ic5c5df350a7eb5291d45cf8e229af575fe9bed57 for featuresets that had it
enabled in queens. Since we won't be backporting the default changes in
tripleoclient back to queens we need to add a special case for these
featuresets in queens.

Change-Id: I9aa449a3198568c25d4dbc838669f493b0e3c103
Closes-Bug: #1771549