Don't trigger repo injection if nothing was built
Currently we have a few possible places for building package in
the job. It requires logic change in order to differentiate
between injection cases.
build-test-packages role can run and not to build anything bc of
various reasons, in this case we need to avoid running
install-built-repo role, that's why we set compressed_gating_repo
to empty string and check in every case if it's empty or not.
If compressed_gating_repo is emptry - we didn't built *last* time
anything, so nothing should run.
See also: https://review.openstack.org/#/c/576277/1

Change-Id: I7d768c5b550984cdd45d8a9cbbf89767e5e1c418