Add option to install additional roles
Add option to quickstart to install additional roles via command
line. Sometimes we need to install external role and we can't
or don't want to create another requirement file for that.
Let's add command line option for that.
The default behaviour will remain the same - installing only
quickstart and extras requirements.
If external or any other requirement is specified, the quickstart
requirements file will be installed automatically,
but not the extras.

Change-Id: Ib0f96f658cf83661c7abae87a5bb77abe12d3f14