Install openstack-heat-agents for Pike subnodes
As described in the related bug below we are removing the tripleo.sh
bootstrap subnodes from ci scripts with the work tracked at [1]

For Queens and Master bootstrap package installation is done with the
tripleo-bootstrap role which doesn't exist on Pike. This adds a
repo_cmd_after to the Pike config installing the required 
openstack-heat-agents otherwise the deployment fails as in the bug.

[1] https://trello.com/c/GvjcnJB2/850-translate-tripleosh-bootstrap-subnodes-into-a-series-of-tasks
Change-Id: Ib95b02767488aaddfe95633d03cb118dd492a700
Closes-Bug: 1785067