Fix delorean-deps url for releases >= rocky
Since Rocky, RDO has split dependencies repo in two different
repos, one for runtime requirement with the URL used in previous
releases and a new one with the packages that are only required
to build packages.
This patch fixes repos to the correct set for rocky and master
in the releases file which have these wrongly set.

Change-Id: I94621e899c1e854d8a762cb05b495b8fd9178ef7