Use mirrored etcd container for openshift CI
We need to use the fedora etcd container for openshift, because we
require the default user to be root[1]. However, this registry is not
mirrored in openstack infra, so we end up getting timeouts. The
fedora container has been manually copied into our docker.io registry
which is mirrored and will not have the same timeout issues.

[1] https://review.openstack.org/576497

Depends-On: I48fcbcfba5e8b41f28e23773e2cc02229f7eee7d
Depends-On: I8e6bc6d46a32a55098991e1916cedebb7e0b5c93
Change-Id: Idfbaf756c49957219ab1f526a2666b43aaddce8b