Fix queens fs016 fs019
In queens, the ceph-ansible workflow does not work with config download.
These jobs should not be using config-download.  This was improperly
converted as part of I881f92e6cef4de58a9731f03669e42bc862964ec. This is
a partial revert of that patch for featureset016 and featureset019.  The
featureset documentation indicates that these are not configured with
config download.

https://docs.openstack.org/tripleo-quickstart/latest/feature-configuration.html

Change-Id: Ib6c3ca7886c5e4adb0883c501bfbb390e15d108f
Closes-Bug: #1790144
Related-Bug: #1789416