Pin openshift-ansible to last known good version
The scenario009 job can easily be broken by changes in
openshift-ansible. There is currently no openstack based testing
there.

In order to keep our job stable (and move it to voting), we need
to prevent these breaking changes from affecting our jobs. We will
need to periodically try to update this pin and fix the issues.

Change-Id: I05457605a1265e7c44f92a883d17cca3e7b0ccf6