Set containerized_undercloud for OpenShift featureset
We are getting most of the behaviors of containerized_undercloud
when using containerized_undercloud=false (default). Namely, the
undercloud is deployed via containers.

However, without this explicity set, we do not include the
containers-prepare-parameter.yaml in the undercloud.conf template[1].

This causes the undercloud containers to never be updated since the
update params are templated to that file.

[1] https://github.com/openstack/tripleo-quickstart-extras/blob/b3794ff03ab841bbbc15a339e35a7ec02b193111/roles/undercloud-deploy/templates/undercloud.conf.j2#L260-L266
Change-Id: Ibe94a58e0083b75e2f4cc904be6c17b0633d0293