Put create repo script into its own tasks file.
This enables repo-setup variables and scripts to be re-generated
without including the entire role during upgrade.

First used in the standalone-upgarde.yml playbook.

The problem is that the second run is prevented when we use the
main.yml task.  As a matter of fact it checks if dlrn_* hash/path
calculation has already happened using a "defined" test on cached
variable and ansible can't undefined variable.

This offers a way to override this limitation by including all
sub-tasks without using main.  This has the benefit of not changing
the repo-setup logic, which may be used/useful for other purposes.

The downside is code duplication.  Another way would be to "fix" the
repo-setup role to be more flexible, but it may have unseen side
effect and should be properly planned.

Change-Id: I6a95aa53f65a720b6f624f6649af7bb43f28860c
Closes-Bug: #1796270